+++ htdocs/inc/xajax/global.xajax.php
@@ -162,139 +162,85 @@ function getFlightTargetInfo($f,$sx1,$sy1,$cx1,$cy1,$p1)
 		}
 	}
 
-	if ($total>0)
-	{
-		if ($p>0)
-		{
-			$res = dbquery("
-			SELECT
-				planet_name,
-				user_nick,
-				user_id,
-				entities.cell_id,
-				planets.id
-
-			FROM 
-				entities
-			INNER JOIN 
-				cells 
-			ON 
-				entities.cell_id = cells.id
-				AND entities.pos=".intval($p)."
-				AND cells.sx=".intval($sx)."
-				AND cells.sy=".intval($sy)."
-				AND cells.cx=".intval($cx)."
-				AND cells.cy=".intval($cy)."
-			INNER JOIN
-				planets
-			ON 
-				entities.id=planets.id
-			LEFT JOIN
-				users
-			ON planet_user_id=user_id				
-				
-			");
-			$out = mysql_num_rows($res);
-			if (mysql_num_rows($res)>0)
-			{
-				$arr=mysql_fetch_array($res);
-				if ($arr['planet_name']!='')
-				{
-					$out = "<b>Planet:</b> ".$arr['planet_name'];
-				}
-				else
-				{
-					$out = "<i>Unbenannter Planet</i>";
-				}
-				if ($arr['user_id']>0)
-				{
-					$out.=" <b>Besitzer:</b> ".$arr['user_nick'];
-					if ($s->getInstance()->user_id==$arr['user_id'] && $arr['user_id']>0)
-					{
-						$out.=' (Eigener Planet)';
-						$objResponse->assign("targetinfo","style.color",'#f00');
-						$launch=false;
-					}
-					else
-					{
-						$objResponse->assign("targetinfo","style.color",'#0f0');
-					}
-				}
-				else
-				{
-					$objResponse->assign("targetinfo","style.color",'#f00');
-					$launch=false;
-				}
-				$objResponse->assign("targetinfo","innerHTML",$out);
-				$objResponse->assign("targetcell","value",$arr['cell_id']);
-				$objResponse->assign("targetplanet","value",$arr['id']);
-			}
-			else
-			{
+	if ($total > 0) {
+		if ($p > 0) {
+			$target = Entity::createFactoryByCoords($sx, $sy, $cx, $cy, $p);
+			if (!$target instanceof Planet) {
 				$objResponse->assign("targetinfo","innerHTML","Hier existiert kein Planet!");
 				$objResponse->assign("targetinfo","style.color",'#f00');
 				$launch=false;
 				$objResponse->assign("targetcell","value",0);
 				$objResponse->assign("targetplanet","value",0);
+			} else {
+				if ($target->name()) {
+					$out = "<b>Planet:</b> ".$target->name();
+				} else {
+					$out = "<i>Unbenannter Planet</i>";
+				}
+
+				if ($target->owner()) {
+					$out.=" <b>Besitzer:</b> " . $target->owner()->nick;
+					if ($s->getInstance()->user_id == $target->owner()->id) {
+						$out.=' (Eigener Planet)';
+						$objResponse->assign("targetinfo","style.color",'#f00');
+						$launch=false;
+					} else {
+						$objResponse->assign("targetinfo","style.color",'#0f0');
+					}
+				} else {
+					$objResponse->assign("targetinfo","style.color",'#f00');
+					$launch=false;
+				}
+
+				$objResponse->assign("targetinfo","innerHTML",$out);
+				$objResponse->assign("targetcell","value",$target->cellId());
+				$objResponse->assign("targetplanet","value",$target->id());
 			}
 
-			// Calc time and distance
-			$nx=$conf['num_of_cells']['p1'];		// Anzahl Zellen Y
-			$ny=$conf['num_of_cells']['p2'];		// Anzahl Zellen X
-			$ae=$conf['cell_length']['v'];			// Länge vom Solsys in AE
-			$np=$conf['num_planets']['p2'];			// Max. Planeten im Solsys
-			$dx = abs(((($sx-1) * $nx) + $cx) - ((($sx1-1) * $nx) + $cx1));
-			$dy = abs(((($sy-1) * $nx) + $cy) - ((($sy1-1) * $nx) + $cy1));
-			$sd = sqrt(pow($dx,2)+pow($dy,2));			// Distanze zwischen den beiden Zellen
-			$sae = $sd * $ae;											// Distance in AE units
-			if ($sx1==$sx && $sy1==$sy && $cx1==$cx && $cy1=$cy)
-				$ps = abs($p-$p1)*$ae/4/$np;				// Planetendistanz wenn sie im selben Solsys sind
-			else
-				$ps = ($ae/2) - (($p)*$ae/4/$np);	// Planetendistanz wenn sie nicht im selben Solsys sind
-			$ssae = $sae + $ps;
-			$timeforflight = $ssae / $speed * 3600;
+			if ($target) {
+				$distanceValue = $target->distanceByCoords($sx1, $sy1, $cx1, $cy1, $p1);
+				$timeforflight = $distanceValue / $speed * 3600;
+				$distance = sprintf('%s AE', nf($distanceValue));
+			} else {
+				$distanceValue = -1;
+				$distance = '-';
+				$timeforflight = null;
+			}
 
 			$objResponse->assign("time","innerHTML",tf($timeforflight));
 			$objResponse->assign("timeforflight","value",$timeforflight);
-			$objResponse->assign("distance","innerHTML",nf($ssae)." AE");
-			if ($ssae > $range)
-			{
+			$objResponse->assign("distance","innerHTML",$distance);
+
+			if ($distanceValue === -1) {
+				$objResponse->assign("distance","style.color","#f00");
+				$launch=false;
+			} elseif ($distanceValue > $range) {
 				$objResponse->assign("distance","style.color","#f00");
 				$objResponse->append("distance","innerHTML"," (zu weit entfernt, ".nf($range)." max)");
 				$launch=false;
-			}
-			else
-			{
+			} else {
 				$objResponse->assign("distance","style.color","#0f0");
 			}
 			$objResponse->assign("speed","innerHTML",round($speed,2)." AE/h");
-
-		}
-		else
-		{
+		} else {
 			$launch=false;
 		}
-	}
-	else
-	{
+	} else {
 		$objResponse->assign("targetinfo","innerHTML","Keine Raketen gewählt!");
 		$launch=false;
 	}
 
-	if ($launch)
-	{
+	if ($launch) {
 		$objResponse->assign("launchbutton","style.color",'#0f0');
 		$objResponse->assign("launchbutton","disabled",false);
-	}
-	else
-	{
+	} else {
 		$objResponse->assign("launchbutton","style.color",'#f00');
 		$objResponse->assign("launchbutton","disabled",true);
 	}
 
 	$objResponse->append("targetinfo","innerHTML",ob_get_contents());
 	ob_end_clean();
-  return $objResponse;
+	return $objResponse;
 }